w3c-reports/12092014/mac-wk2/runned-tests/cors/basic-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8801/cors/resources/cors-makeheader.py?get_value=hest_er_best&origin=none&2. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8801/cors/resources/cors-makeheader.py?get_value=hest_er_best&origin=none&3. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?get_value=hest_er_best&origin=none&1. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8801/cors/resources/cors-makeheader.py?get_value=hest_er_best&3. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?get_value=hest_er_best&1. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.

PASS Same domain basic usage 
FAIL Cross domain basic usage assert_true: Got CORS-allowed response expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/cors/basic.htm:36:28
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS Same domain different port 
FAIL Cross domain different port assert_true: Got CORS-allowed response expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/cors/basic.htm:36:28
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
